restoring the runsummaries. a complicated thing between dropbox and github happened that caused it to be deleted: Take note.
my dropbox folder is available on both my mac (com1) and desktop (com2). This repo is available on dropbox, and therefore, on both com1 and com2. But really, this dropbox version should be thought of as the working copy for com1, NOT com2. We will call this working copy WC1.

Now, this repo also has a local non-dropbox working copy on com2. We will call this working copy WC2.

The problem occured when i edited a file (runsummareis.m) that was in WC1 on com2. but then from com2, i committed and pushed WC2 (which had the unedited file), and pulled this change onto com1's working copy, WC1. In this process, i overwrote all my changes to the file.

So really, word of advice:
1. whenever you make a change, puch it to the repo immediately, while it is still fresh.
2. Never change anything from com2 on the dropbox folder, and then expect that change to still be there when you push WC2 onto the remote repo.

Actually I think this problem occurred because I created the file on both WC1 and WC2, but did not commit either. I made changes to the one in WC1, but not WC2, and then committed the one in WC2, and then pulled this to overwrite the one in WC1, losing that copy in WC1.